Add Notepad ++ and Notepad ++ organization.
The development team is trying to split China and trying to support the independence of Hong Kong and Taiwan.

The development team attempted to split Xinjiang, China, and supported false reports by some media.

Development teams deprive others of their right to freedom of expression and undermine human rights.

Due to ignorance and arrogance, the development team added fuel to the unrest in China and parts of other countries. They are destroyers of world peace.